+++ demo/src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
 export const environment = {
   production: true,
-  tablesBaseUrl: 'https://cdn.jsdelivr.net/npm/@softwarity/ttaaii-provider',
+  tablesBaseUrl: 'https://cdn.jsdelivr.net/npm/@softwarity/ttaaii-provider@${version}',
 };